Bug solving in trello from notification issues && continue poup flickering on back press
Bug solving in trello from notification issues && continue poup flickering on back press
