+++ app/screens/home/Childgrowth.tsx
@@ -123,10 +123,10 @@ const {width,height}= Dimensions.get('window');
         visible={modalVisible}
         onRequestClose={() => {
           // Alert.alert('Modal has been closed.');
-          setModalVisible(false);
+          // setModalVisible(false);
         }}
         onDismiss={() => {
-          setModalVisible(false);
+          // setModalVisible(false);
         }}>
         <PopupOverlay>
           <ModalPopupContainer>
