+++ app/screens/home/HealthCheckups.tsx
@@ -180,10 +180,10 @@ const HealthCheckups = ({navigation,route}: Props) => {
         visible={modalVisible}
         onRequestClose={() => {
           // Alert.alert('Modal has been closed.');
-          setModalVisible(false);
+          // setModalVisible(false);
         }}
         onDismiss={() => {
-          setModalVisible(false);
+          // setModalVisible(false);
         }}>
         <PopupOverlay>
           <ModalPopupContainer>
